Change to having a generator per socket.
This enables easier and more performant response parsing as we can now
rely on protocol ordering properties (e.g., FIFO), which we couldn't do
with a single global generator.

We also clean up the reference counting, doing it at the generator level
which is a little clearner and easier to understand.

Finally, for now we disable the memcache protocol as it hasn't been
ported to the new design.