Add support for calculating queueing at load generator
We add support for write callbacks to the Socket abstraction, this
allows us to timestamp when an application packet is actually copied to
the kernel. We use this to give us a poor mans queueing delay (generator
side only) for every protocol.

We also clean up the output format given the new information complexity.